+++ benchmark/benchmarks/react/1000fields/index.ts
@@ -6,7 +6,7 @@ import {
   endBenchmarkLog,
   getCycleResult,
   startBenchmarkLog,
-} from '../../../benchmarkManager';
+} from '../../benchmarkManager';
 
 // Files to run the Benchmark on
 import agileCollection from './bench/agilets/collection';
