+++ benchmark/benchmarks/typescript/defineConfig/index.ts
@@ -5,7 +5,7 @@ import {
   endBenchmarkLog,
   getCycleResult,
   startBenchmarkLog,
-} from '../../../benchmarkManager';
+} from '../../benchmarkManager';
 
 // Files to run the Benchmark on
 import * as referencer from './bench/referencer';
