+++ benchmark/package.json
@@ -7,7 +7,7 @@
   "homepage": "https://agile-ts.org/",
   "description": "Benchmark Tests",
   "scripts": {
-    "test": "node -r esbuild-register run.ts",
+    "test": "node -r esbuild-register runtime/run.ts",
     "test:counter": "yarn test ./benchmarks/react/counter",
     "test:1000fields": "yarn test ./benchmarks/react/1000fields",
     "test:computed": "yarn test ./benchmarks/react/computed",
